updates from pr
