+++ src/app/container/container.component.html
@@ -190,7 +190,7 @@
       <div *ngIf="tool?.topicId !== null else noTopicId" id='discourse-comments'></div>
       <ng-template #noTopicId>
         <mat-card class="alert alert-warning" role="alert">
-          <mat-icon>info</mat-icon> No Discourse topic exists for this tool. Contact a Dockstore curator to create a topic.
+          <mat-icon>info</mat-icon> No Discourse topic exists for this tool.
         </mat-card>
       </ng-template>
     </div>
