+++ src/app/workflow/workflow.component.html
@@ -195,7 +195,7 @@
     <div *ngIf="workflow?.topicId !== null else noTopicId" id='discourse-comments'></div>
     <ng-template #noTopicId>
       <mat-card class="alert alert-warning" role="alert">
-        <mat-icon>info</mat-icon> No Discourse topic exists for this workflow. Contact a Dockstore curator to create a topic.
+        <mat-icon>info</mat-icon> No Discourse topic exists for this workflow.
       </mat-card>
     </ng-template>
   </div>
